+++ src/modules/rules-presets.ts
@@ -8,6 +8,9 @@ function getStdLintRules() {
     // 作者、期刊、年、期、卷、页 -> 判断语言 -> 作者大小写 -> 匹配缩写 -> 匹配地点 -> 格式化日期 -> 格式化DOI
     const rules = [];
     // getPref("isEnableOtherFields") ? rules.push(new Rules.UpdateMetadata({ mode: "blank" })) : "";
+    getPref("noDuplicationItems") ? rules.push(new Rules.NoDuplicatItem({})) : "skip";
+    getPref("checkWebpage") ? rules.push(new Rules.NoWebPageItem({})) : "skip";
+    getPref("noPreprintJournalArticle") ? rules.push(new Rules.NoPreprintJournalArticle({})) : "skip";
     getPref("lang") ? rules.push(new Rules.UpdateItemLanguage({})) : "";
     getPref("creatorsCase") ? rules.push(new Rules.CapitalizeCreators({})) : "";
     getPref("titleSentenceCase") ? rules.push(new Rules.TitleSentenceCase({})) : "";
@@ -24,8 +27,6 @@ function getStdLintRules() {
 
 function getNewItemLintRules() {
     const rules = [];
-    getPref("checkWebpage") ? rules.push(new Rules.NoWebPageItem({})) : "skip";
-    getPref("noDuplicationItems") ? rules.push(new Rules.NoDuplicatItem({})) : "skip";
     getPref("lint.onAdded") ? rules.push(getStdLintRules()) : "";
     return rules.flat();
 }