+++ src/modules/rules/index.ts
@@ -5,6 +5,7 @@ import DateISO from "./field-date-iso";
 import RemoveDOIPrefix from "./field-doi-no-prefix";
 import UpdateJournalAbbr from "./field-journalAbbr";
 import UpdateItemLanguage from "./field-language";
+import PagesConnector from "./field-pages";
 import UpdateUniversityPlace from "./field-place";
 import UpdatePublicationTitle from "./field-publication";
 import ThesisType from "./field-thesisType";
@@ -38,4 +39,5 @@ export default {
     ThesisType,
     University,
     NoPreprintJournalArticle,
+    PagesConnector,
 };