+++ scripts/build.mjs
@@ -149,7 +149,7 @@ export const esbuildOptions = {
     target: "firefox102",
     outfile: path.join(buildDir, `addon/content/scripts/${config.addonRef}.js`),
     // Don't turn minify on
-    minify: env.NODE_ENV === "production",
+    // minify: env.NODE_ENV === "production",
 };
 
 export async function main() {